--- src/process.py
@@ -1,120 +1,0 @@
-import logging
-from typing import TYPE_CHECKING
-
-import src.globals as g
-
-from .constants import BRANCH_NAME_PREFIX
-from .models import PartialGitHubPullRequestEvent, PublishInfo
-from .utils import (
-    comment_issue,
-    commit_and_push,
-    create_pull_request,
-    extract_issue_number_from_ref,
-    extract_publish_info_from_issue,
-    get_pull_requests_by_label,
-    get_type_by_labels,
-    get_type_by_title,
-    resolve_conflict_pull_requests,
-    run_shell_command,
-    should_skip_plugin_test,
-)
-
-if TYPE_CHECKING:
-    from github.Repository import Repository
-
-
-def process_pull_request_event(repo: "Repository"):
-    """处理 Pull Request 事件"""
-    event = PartialGitHubPullRequestEvent.parse_file(g.settings.github_event_path)
-    logging.info(f"当前事件: {event.json()}")
-
-    # 因为合并拉取请求只会触发 closed 事件
-    # 其他事件均对发布流程无影响
-    if event.action != "closed":
-        logging.info("事件不是关闭拉取请求，已跳过")
-        return
-
-    pull_request = repo.get_pull(event.pull_request.number)
-
-    # 只处理支持标签的拉取请求
-    publish_type = get_type_by_labels(pull_request.labels)
-    if not publish_type:
-        logging.info("拉取请求与发布无关，已跳过")
-        return
-
-    ref = pull_request.head.ref
-    related_issue_number = extract_issue_number_from_ref(ref)
-    if not related_issue_number:
-        logging.error("无法获取相关的议题编号")
-        return
-
-    issue = repo.get_issue(related_issue_number)
-    if issue.state == "open":
-        logging.info(f"正在关闭议题 #{related_issue_number}")
-        issue.edit(state="closed")
-    logging.info(f"议题 #{related_issue_number} 已关闭")
-
-    try:
-        run_shell_command(["git", "push", "origin", "--delete", ref])
-        logging.info(f"已删除对应分支")
-    except:
-        logging.info("对应分支不存在或已删除")
-
-    if pull_request.merged:
-        logging.info("发布的拉取请求已合并，准备更新拉取请求的提交")
-        pull_requests = get_pull_requests_by_label(repo, publish_type.value)
-        resolve_conflict_pull_requests(pull_requests, repo)
-    else:
-        logging.info("发布的拉取请求未合并，已跳过")
-
-
-def process_issues_event(repo: "Repository", issue_number: int):
-    """处理议题"""
-    issue = repo.get_issue(issue_number)
-
-    if issue.pull_request:
-        logging.info("评论在拉取请求下，已跳过")
-        return
-
-    if issue.state != "open":
-        logging.info("议题未开启，已跳过")
-        return
-
-    publish_type = get_type_by_title(issue.title)
-    if not publish_type:
-        logging.info("议题与发布无关，已跳过")
-        return
-
-    # 自动给议题添加标签
-    issue.edit(labels=[publish_type.value])
-
-    # 是否需要跳过插件测试
-    g.skip_plugin_test = should_skip_plugin_test(issue)
-
-    # 检查是否满足发布要求
-    # 仅在通过检查的情况下创建拉取请求
-    info = extract_publish_info_from_issue(issue, publish_type)
-    if isinstance(info, PublishInfo):
-        # 拉取请求与议题的标题
-        title = f"{info.get_type().value}: {info.name}"
-        # 修改议题标题
-        if issue.title != title:
-            issue.edit(title=title)
-            logging.info(f"议题标题已修改为 {title}")
-        # 创建新分支
-        # 命名示例 publish/issue123
-        branch_name = f"{BRANCH_NAME_PREFIX}{issue.number}"
-        run_shell_command(["git", "switch", "-C", branch_name])
-        # 更新文件并提交更改
-        info.update_file()
-        commit_and_push(info, branch_name, issue.number)
-        # 创建拉取请求
-        create_pull_request(
-            repo, info, g.settings.input_config.base, branch_name, issue.number, title
-        )
-        message = info.validation_message
-    else:
-        message = info.message
-        logging.info("发布没通过检查，暂不创建拉取请求")
-
-    comment_issue(issue, message)
